Better mpv invocation behavior
elfeed-tube-mpv-play is now just elfeed-tube-mpv

It can be called in the following ways:

In an elfeed-show buffer:
- Invoked via M-x elfeed-tube-mpv or a custom keybinding
- By clicking on a segment in the transcript
- By pressing RET on a segment in the transcript

In an elfeed-search buffer:
- Invoked via M-x elfeed-tube-mpv or a custom keybinding. It acts on the entry
at point.

Its behavior is as follows:

Called directly: Play the video while "connected to Elfeed"
Called directly with point on the transcript: Play at point, or seek to point
Called by clicking/pressing RET on transcript: Play at point, or seek to point
Called on another video when already playing one: add to playlist
Called with a prefix argument: Start a new instance of mpv that is "unconnected to Elfeed."

Still testing how this goes.

New user option elfeed-tube-mpv-options