Simultaneous fetching of description and captions
Using an aio-select object. I'm not sure how much this speeds up the fetching,
and if it's worth the additional effort of understanding the new code. Need to test.

Add support for chapters in transcript, imenu navigation

New option 'chapters in elfeed-tube-fields to fetch/display video chapters. It
is enabled by default.

Add demo video of chapters and imenu navigation.