Add elfeed-tube-fill.el.
This module allows back-filling of Youtube feeds in the Elfeed database.
Back-filling a feed fetches all entries for the corresponding YouTube channel or
playlist and adds them to the Elfeed database. Youtube RSS feeds generally
contain only the latest 15 entries.

Call `elfeed-tube-fill-feeds' in an Elfeed search or entry buffer to
back-fill entries for the corresponding feed(s). You can select a region to
specify a list of feeds.